requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
tavily-python>=0.5.0

pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
typing-extensions>=4.11.0

openai>=1.35.0
anthropic>=0.34.0

# Dev / logging niceties (optional, but harmless)
rich>=13.7.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
